Move AppMenu offset calculator function to separate module
