+++ src/components/AppMenu/AppMenu.tsx
@@ -2,12 +2,13 @@ import * as React from 'react';
 import { Row } from '../Row';
 import { createTransformer } from './createTransformer';
 import { useRefNormalizer } from '../../hooks/useRefNormalizer';
-import { AppMenuItem, itemGutter, itemWidth } from './AppMenuItem';
+import { AppMenuItem, itemHeight, itemWidth } from './AppMenuItem';
 import { Content } from '../../state/Content';
-import styled, { DefaultTheme, ThemeContext } from 'styled-components';
+import styled, { ThemeContext } from 'styled-components';
 import { activationTransition } from '../../css/transitions';
 import { math } from 'polished';
 import { useSpatialIndex } from '../../lib/spatial/useSpatialIndex';
+import { getOffsetForActiveItem } from './getOffsetForActiveItem';
 
 export type AppMenuProps = React.HTMLAttributes<HTMLDivElement> & {
   items?: Content[];
@@ -17,12 +18,9 @@ export const AppMenu = React.forwardRef<HTMLDivElement, AppMenuProps>(
   ({ items = [], ...props }, ref) => {
     const index = useSpatialIndex(useRefNormalizer(ref));
     const theme = React.useContext(ThemeContext);
+    const offset = getOffsetForActiveItem(theme, index);
     return (
-      <SlidingRow
-        ref={ref}
-        transform={`translate(${getDistanceToActiveItem(theme, index)})`}
-        {...props}
-      >
+      <SlidingRow ref={ref} transform={`translate(${offset})`} {...props}>
         {items.map((itemProps, index) => (
           <AppMenuItem key={index} activate={index === 0} {...itemProps} />
         ))}
@@ -31,20 +29,8 @@ export const AppMenu = React.forwardRef<HTMLDivElement, AppMenuProps>(
   }
 );
 
-const getDistanceToActiveItem = (theme: DefaultTheme, index: number) => {
-  if (index === -1) {
-    return 0;
-  }
-  const normalGutters = math(
-    `max(0, ${index - 1}) * ${itemGutter(theme, false)}`
-  );
-  const largeGutter = index > 0 ? itemGutter(theme, true) : 0;
-  return math(
-    `-(${index} * ${itemWidth(theme)} + ${normalGutters} + ${largeGutter})`
-  );
-};
-
 const SlidingRow = styled(createTransformer(Row))`
   ${activationTransition('transform')};
   margin-left: ${props => math(`${itemWidth(props.theme)} * 1`)};
+  height: ${props => itemHeight(props.theme, false)};
 `;
